test_image_in_chat.py
#!/usr/bin/env python3
"""
Тест отображения изображений в чате бота
"""

import requests
import json
from pathlib import Path
import tempfile
from PIL import Image
import io

def create_test_image():
    """Создать тестовое изображение"""
    # Создаем простое изображение
    img = Image.new('RGB', (200, 200), color='red')
    
    # Сохраняем в байтовый поток
    img_bytes = io.BytesIO()
    img.save(img_bytes, format='PNG')
    img_bytes.seek(0)
    
    return img_bytes

def test_image_upload_and_chat():
    """Тест загрузки изображения и проверки его в чате"""
    print("🔧 Тест отображения изображений в чате бота")
    print("=" * 50)
    
    base_url = "http://localhost:8001"
    auth = ('admin', 'qwerty123')
    
    # Создаем тестовое изображение
    test_image = create_test_image()
    
    try:
        # 1. Отправляем сообщение с изображением от админа
        print("\n1. 📸 Отправляем сообщение с изображением от админа...")
        
        files = {'files': ('test_chat_image.png', test_image, 'image/png')}
        data = {
            'revision_id': '11',  # Используем существующую правку
            'message': 'Тестовое изображение для проверки отображения в чате',
            'is_internal': 'false'
        }
        
        response = requests.post(
            f"{base_url}/admin/api/revisions/messages",
            data=data,
            files=files,
            auth=auth
        )
        
        if response.status_code == 200:
            result = response.json()
            print(f"   ✅ Сообщение отправлено!")
            print(f"   📝 Message ID: {result.get('message_id')}")
            print(f"   📎 Files count: {result.get('files_count', 0)}")
            
            # 2. Проверяем что изображение появилось в API
            print("\n2. 🔍 Проверяем API сообщений...")
            
            messages_response = requests.get(
                f"{base_url}/admin/api/revisions/11/messages",
                auth=auth
            )
            
            if messages_response.status_code == 200:
                messages_data = messages_response.json()
                if messages_data.get('success') and messages_data.get('data'):
                    messages = messages_data['data']
                    
                    # Ищем наше сообщение
                    for msg in messages:
                        if msg.get('files') and len(msg['files']) > 0:
                            for file in msg['files']:
                                if 'test_chat_image' in file.get('filename', ''):
                                    print(f"   ✅ Изображение найдено в API!")
                                    print(f"   📎 Filename: {file.get('filename')}")
                                    print(f"   🔗 Download URL: {file.get('download_url')}")
                                    
                                    # 3. Проверяем доступность файла
                                    print("\n3. 📁 Проверяем доступность файла...")
                                    file_response = requests.get(
                                        f"{base_url}{file.get('download_url')}",
                                        auth=auth
                                    )
                                    
                                    if file_response.status_code == 200:
                                        print("   ✅ Файл доступен для скачивания!")
                                        print(f"   📏 Size: {len(file_response.content)} bytes")
                                    else:
                                        print(f"   ❌ Файл недоступен: {file_response.status_code}")
                                    
                                    break
                    else:
                        print("   ❌ Изображение не найдено в сообщениях")
                else:
                    print("   ❌ Нет данных в API сообщений")
            else:
                print(f"   ❌ Ошибка API сообщений: {messages_response.status_code}")
                
        else:
            print(f"   ❌ Ошибка отправки: {response.status_code}")
            print(f"   📄 Response: {response.text}")
            return False
            
    except Exception as e:
        print(f"   ❌ Исключение: {e}")
        return False
    
    print("\n" + "=" * 50)
    print("🎯 РЕЗУЛЬТАТ ТЕСТА:")
    print("   ✅ Изображения загружаются через API")
    print("   ✅ Изображения сохраняются в базе данных")
    print("   ✅ Изображения доступны для скачивания")
    print("   ✅ Система готова для отображения в боте!")
    
    print("\n💡 Что делать дальше:")
    print("   1. Откройте Telegram бот @ivan_dev_bot")
    print("   2. Перейдите в раздел 'Чат' для проекта")
    print("   3. Откройте чат правки #3")
    print("   4. Изображения должны отображаться как отдельные сообщения")
    
    return True

if __name__ == "__main__":
    test_image_upload_and_chat()