requirements.txt

# Telegram Bot
python-telegram-bot>=20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Web Framework for Admin Panel
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6

# AI Services
openai>=1.0.0
httpx>=0.25.0

# Speech Recognition
SpeechRecognition>=3.10.0
pydub>=0.25.1

# File Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
Pillow>=10.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
APScheduler>=3.10.0
redis>=4.5.0
celery>=5.3.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Additional
aiofiles>=23.0.0
python-dateutil>=2.8.0
pytz>=2023.0